+++ braintree/exceptions/timeout_error.py
@@ -1,7 +1,0 @@
-from braintree.exceptions.braintree_error import BraintreeError
-
-class TimeoutError(BraintreeError):
-    """
-    Raised when a Timeout occurs
-    """
-    pass